refactor(robot): rename robots.py to robot_controller.py
