+++ plb/engine/controller/__init__.py
@@ -1,1 +1,2 @@
+from .controller import Controller
 from .primitive_controller import Primitives
