+++ tests/unit/test_robot_grad.py
@@ -2,7 +2,7 @@ import taichi
 import torch
 
 from plb.engine.controller.primitive_controller import Primitives
-from plb.engine.controller.robots import RobotsController
+from plb.engine.controller.robot_controller import RobotsController
 from plb.urdfpy import DiffRobot
 
 taichi.init()
